move ContiguousStorage implementation
